Remove the loader's evaluator, import.meta and surgical-invalidation seams
The Loader again evaluates every module in its own realm and hands
`import.meta` `{ url, loader }` directly. `invalidateModule`,
`isShimmedModule` and the record of which fetches answered with a shim go
with them; nothing outside their own tests called them.

What stays is the guards a realm-mapping change reaches on its own: a
fetch whose record was discarded mid-flight neither writes back nor evicts
its replacement, an import whose walk resolved into an emptied map
re-enters it, and a dependency walk taken over a still-fetching module is
not memoized. Their tests move beside the other mapping-change tests, with
the mapping change as the trigger.
